Modified SymbolTable structure.
Each variable stored in Symbol Table was earlier mapped to std::string, i.e the var's data.
This commit modifies this structure. Each variable is now mapped to a struct that stores
more comprehensive details about the variable, including it's type and also adds the hooks
required to allow user to declare an array.